LPD-23201 Move ObjectFieldDBTypeUtil to Object Info API module
