+++ modules/apps/object/object-info-api/src/main/java/com/liferay/object/info/util/ObjectFieldDBTypeUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.web.internal.util;
+package com.liferay.object.info.util;
 
 import com.liferay.info.field.type.BooleanInfoFieldType;
 import com.liferay.info.field.type.DateInfoFieldType;
