+++ modules/apps/object/object-info-api/src/main/resources/com/liferay/object/info/util/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
